pyproject.toml
[tool.poetry]
name = "microservices-benchmark"
version = "0.1.0"
description = ""
authors = ["Aryan Kumar <kumarayan86@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
grpcio = "^1.67.1"
grpcio-tools = "^1.67.1"
protobuf = "^5.28.3"
pydantic = "^2.9.2"
prometheus-client = "^0.21.0"
locust = "^2.32.1"
pytest = "^8.3.3"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"
watchdog = "^5.0.3"
grpc-interceptor = "^0.15.4"
gevent = "^24.10.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
